gpgsig -----BEGIN PGP SIGNATURE-----  Version: GnuPG v1.4.11 (GNU/Linux)
 iQEcBAABAgAGBQJRLFLYAAoJEHQiPpGmrK+iqvMIALY3l/zFv6Xia4zZ3laKZvPX
 JeVokfoQs2ErCyMr+8pvBMNOXgdjX7pP8sArDGtbsmC6AFMxWmXzDp9QcNvfmy6v
 X9hDPwS8DOznc+3nY0G8DGIrFxA8y/Dm4/ViUs95RhCFkcEN5tElImlJwdd9Ifs6
 55lb1L2Gprk6Q6MBUG/jp9TTuae+5yAcO6O69ZZ/H3Rq7VgQVsOrL4MQsSmvO2s+
 PXb6xdTX1SAuYSu/kNesZthqrBZp1YYs5O4g4oR1vqS0h2NnVfEkSpJCX5FP6qJ7
 LE7dzegYfWKSRsJAGj3ulrHZqwYaCuxXt//kLurc3qSIItTQ5B6+vRKzq1VyrT8=
 =gsYD
 -----END PGP SIGNATURE-----

Fix repl handling of blank line

Since the move to the nodeREPL package, input lines to be evaluated are
now wrapped in parentheses; that is:

    'foo'

would become:

    ('foo'
    )

The old way of detecting empty lines was to see if the input string was
either totally empty, or whitespace-only. The addition of these
parentheses breaks that.

In order to fix this, we simply tweak the regex a little to ignore these
added parentheses if they're present. As an added bonus, the regex
sh